added OpeningHoursVIew.swift
